% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_CODEXCOV.R
\name{qcObj1}
\alias{qcObj1}
\title{Function Dexcription}
\usage{
qcObj1(Y, sampname, chr, ref, mapp, gc, cov_thresh, length_thresh, mapp_thresh,
  gc_thresh)
}
\description{
Function description.
}
\examples{
coverageObj1
}
