% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_CODEX.R
\name{segment1}
\alias{segment1}
\title{Function Dexcription}
\usage{
segment1(Y_qc, Yhat, optK, K, sampname_qc, ref_qc, chr, lmax, mode)
}
\description{
Function description.
}
\examples{
coverageObj1
}
